+++ lib/open-build-service-api.rb
@@ -21,9 +21,11 @@ require_relative 'helpers/binary_helper'
 require_relative 'collections/abstract_collection'
 require_relative 'collections/binaries_collection'
 require_relative 'collections/packages_collection'
+require_relative 'collections/sources_collection'
 
 # Models
 require_relative 'models/binary'
 require_relative 'models/package'
 require_relative 'models/project'
 require_relative 'models/repository'
+require_relative 'models/source'