+++ lib/models/abstract_model.rb
@@ -1,0 +1,4 @@
+module OpenBuildServiceAPI
+  class AbstractModel
+  end
+end